+++ example/shape/hexagon.py
@@ -4,5 +4,5 @@ from cadqueryhelper import shape
 result = shape.regular_polygon(sides=6)
 cq.exporters.export(result,'out/hexagon.stl')
 
-if part.metadata:
-    print(result.metadata)
+#if part.metadata:
+#    print(result.metadata)