+++ src/cadqueryhelper/shape/regularPolygon.py
@@ -11,21 +11,21 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import cqmore as cqm
-from cqmore.polygon import regularPolygon
+import cadquery as cq
 
 def regular_polygon(radius=5, sides=6, height=5):
-    diameter = radius*2
-    work = (
-        cqm.Workplane()
-        .makePolygon(regularPolygon(nSides = sides, radius = radius))
+    polygon = (
+        cq.Workplane("XY")
+        .polygon(sides, radius)
     )
 
     if height:
-        work = work.extrude(height)
+        polygon = (
+            polygon
+            .extrude(height)
+            .translate((0,0,-1*(height/2)))
+        )
 
-    work = work.translate((0,0,-1*(height/2)))
-
-    meta = {'type':'regularPolygon','height':height, 'length':diameter, 'width':diameter}
-    work.metadata = meta
-    return work
+    meta = {'type':'regularPolygon','height':height, 'length':radius*2, 'width':radius*2}
+    polygon.metadata = meta
+    return polygon